+++ backend/DTOs/ScoreHistoryEntryDto.cs
@@ -3,6 +3,6 @@ namespace DailyChallenges.DTOs
     public class ScoreHistoryEntryDto
     {
         public string ScoringDay { get; set; } = string.Empty;
-        public int? ScoreValue { get; set; }
+        public double? ScoreValue { get; set; }
     }
 }
